Fixed segmentation fault on wxDynamicCast caused by use of int instead of size_t data type for the string len of second parameter.

+++ tools/source_maker/templates/functions/wxdynamiccast.php
@@ -3,7 +3,7 @@ PHP_FUNCTION(php_wxDynamicCast)
 {
     zval* object;
     char* object_type;
-    int object_type_len = 0;
+    size_t object_type_len = 0;
     void* native_object = 0;
 
     char parse_parameters[] = "z!s!";
